pyproject.toml
[tool.poetry]
name = "instagram-reels-scroller"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.50.0"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
